MP Fixes
Okay, had to change the mech spawning, the box mechanic, and a number of smaller things (i.e. hints) to be MP compatible, but all of them seem to be working fine.

Only thing I'd like to still do is add more intel items around and check/add more respawns (fixable in game by zeus)

The enemies are also kinda braindead due to the low light, but I'm not sure how we can fix that. :o